@@ -1,6 +1,6 @@
 module github.com/getlantern/lantern-client
 
-go 1.21.4
+go 1.22.0
 
 // replace github.com/getlantern/flashlight/v7 => ../flashlight
 
