@@ -3,9 +3,8 @@ package app
 import (
 	"context"
 	"fmt"
+	"io"
 	"io/fs"
-	"io/ioutil"
-	"math/rand"
 	"net"
 	"net/http"
 	"os"
@@ -30,22 +29,21 @@ import (
 	"github.com/getlantern/flashlight/v7/logging"
 	"github.com/getlantern/flashlight/v7/ops"
 	"github.com/getlantern/flashlight/v7/otel"
-	"github.com/getlantern/flashlight/v7/pro"
-	proclient "github.com/getlantern/flashlight/v7/pro/client"
 	"github.com/getlantern/flashlight/v7/stats"
 	"github.com/getlantern/golog"
 	"github.com/getlantern/i18n"
 	"github.com/getlantern/memhelper"
 	notify "github.com/getlantern/notifier"
 	"github.com/getlantern/profiling"
-	"github.com/getlantern/trafficlog-flashlight/tlproc"
 
 	"github.com/getlantern/lantern-client/desktop/analytics"
 	"github.com/getlantern/lantern-client/desktop/autoupdate"
 	uicommon "github.com/getlantern/lantern-client/desktop/common"
 	"github.com/getlantern/lantern-client/desktop/features"
 	"github.com/getlantern/lantern-client/desktop/notifier"
+	"github.com/getlantern/lantern-client/desktop/settings"
 	"github.com/getlantern/lantern-client/desktop/ws"
+	proclient "github.com/getlantern/lantern-client/internalsdk/pro"
 )
 
 var (
@@ -57,8 +55,6 @@ var (
 func init() {
 	autoupdate.Version = ApplicationVersion
 	autoupdate.PublicKey = []byte(packagePublicKey)
-
-	rand.Seed(time.Now().UnixNano())
 }
 
 // App is the core of the Lantern desktop application, in the form of a library.
@@ -71,7 +67,7 @@ type App struct {
 	configDir        string
 	exited           eventual.Value
 	analyticsSession analytics.Session
-	settings         *Settings
+	settings         *settings.Settings
 	statsTracker     *statsTracker
 
 	muExitFuncs sync.RWMutex
@@ -83,28 +79,13 @@ type App struct {
 
 	flashlight *flashlight.Flashlight
 
-	// If both the trafficLogLock and proxiesLock are needed, the trafficLogLock should be obtained
-	// first. Keeping the order consistent avoids deadlocking.
-
-	// Log of network traffic to and from the proxies. Used to attach packet capture files to
-	// reported issues. Nil if traffic logging is not enabled.
-	trafficLog     *tlproc.TrafficLogProcess
-	trafficLogLock sync.RWMutex
-
-	// Also protected by trafficLogLock.
-	captureSaveDuration time.Duration
-
-	// proxies are tracked by the application solely for data collection purposes. This value should
-	// not be changed, except by Flashlight.onProxiesUpdate. State-changing methods on the dialers
-	// should not be called. In short, this slice and its elements should be treated as read-only.
-	proxies     []bandit.Dialer
-	proxiesLock sync.RWMutex
-
 	issueReporter *issueReporter
-	proClient     *proclient.Client
-	referralCode  string
-	selectedTab   Tab
-	stats         *stats.Stats
+
+	proClient proclient.ProClient
+
+	referralCode string
+	selectedTab  Tab
+	stats        *stats.Stats
 
 	connectionStatusCallbacks []func(isConnected bool)
 	_sysproxyOff              func() error
@@ -117,7 +98,7 @@ type App struct {
 }
 
 // NewApp creates a new desktop app that initializes the app and acts as a moderator between all desktop components.
-func NewApp(flags flashlight.Flags, configDir string, proClient *proclient.Client, settings *Settings) *App {
+func NewApp(flags flashlight.Flags, configDir string, proClient proclient.ProClient, settings *settings.Settings) *App {
 	analyticsSession := newAnalyticsSession(settings)
 	app := &App{
 		configDir:                 configDir,
@@ -135,19 +116,19 @@ func NewApp(flags flashlight.Flags, configDir string, proClient *proclient.Clien
 	golog.OnFatal(app.exitOnFatal)
 
 	app.AddExitFunc("stopping analytics", app.analyticsSession.End)
-	pro.OnProStatusChange(func(isPro bool, _ bool) {
+	onProStatusChange(func(isPro bool) {
 		app.statsTracker.SetIsPro(isPro)
 	})
 
 	log.Debugf("Using configdir: %v", configDir)
 
-	app.issueReporter = newIssueReporter(app.settings, app.getCapturedPackets, app.getProxies)
+	app.issueReporter = newIssueReporter(app)
 	app.translations.Set(os.DirFS("locale/translation"))
 
 	return app
 }
 
-func newAnalyticsSession(settings *Settings) analytics.Session {
+func newAnalyticsSession(settings *settings.Settings) analytics.Session {
 	if settings.IsAutoReport() {
 		session := analytics.Start(settings.GetDeviceID(), ApplicationVersion)
 		go func() {
@@ -194,7 +175,7 @@ func (app *App) Run(isMain bool) {
 
 		listenAddr := app.Flags.Addr
 		if listenAddr == "" {
-			listenAddr = app.settings.getString(SNAddr)
+			listenAddr = app.settings.GetAddr()
 		}
 		if listenAddr == "" {
 			listenAddr = defaultHTTPProxyAddress
@@ -202,7 +183,7 @@ func (app *App) Run(isMain bool) {
 
 		socksAddr := app.Flags.SocksAddr
 		if socksAddr == "" {
-			socksAddr = app.settings.getString(SNSOCKSAddr)
+			socksAddr = app.settings.GetSOCKSAddr()
 		}
 		if socksAddr == "" {
 			socksAddr = defaultSOCKSProxyAddress
@@ -239,8 +220,8 @@ func (app *App) Run(isMain bool) {
 			RevisionDate,
 			app.configDir,
 			app.Flags.VPN,
-			func() bool { return app.settings.getBool(SNDisconnected) }, // check whether we're disconnected
-			func() bool { return false },                                // on desktop, we do not allow private hosts
+			func() bool { return app.settings.GetDisconnected() }, // check whether we're disconnected
+			func() bool { return false },                          // on desktop, we do not allow private hosts
 			app.settings.IsAutoReport,
 			app.Flags.AsMap(),
 			app.settings,
@@ -260,11 +241,11 @@ func (app *App) Run(isMain bool) {
 		app.beforeStart(listenAddr)
 
 		chProStatusChanged := make(chan bool, 1)
-		pro.OnProStatusChange(func(isPro bool, _ bool) {
+		onProStatusChange(func(isPro bool) {
 			chProStatusChanged <- isPro
 		})
 		chUserChanged := make(chan bool, 1)
-		app.settings.OnChange(SNUserID, func(v interface{}) {
+		app.settings.OnChange(settings.SNUserID, func(v interface{}) {
 			chUserChanged <- true
 		})
 		app.startFeaturesService(geolookup.OnRefresh(), chUserChanged, chProStatusChanged, app.chGlobalConfigChanged)
@@ -280,7 +261,7 @@ func (app *App) Run(isMain bool) {
 						Title:      i18n.T("BACKEND_CONFIG_SAVE_ERROR_TITLE"),
 						Message:    i18n.T("BACKEND_CONFIG_SAVE_ERROR_MESSAGE", i18n.T(translationAppName)),
 						ClickLabel: i18n.T("BACKEND_CLICK_LABEL_GOT_IT"),
-						IconURL:    app.AddToken("/img/lantern_logo.png"),
+						IconURL:    "/img/lantern_logo.png",
 					}
 					_ = notifier.ShowNotification(note, "alert-prompt")
 				})
@@ -319,8 +300,6 @@ func (app *App) checkEnabledFeatures() {
 
 	log.Debugf("Starting enabled features: %v", enabledFeatures)
 	//go app.startReplicaIfNecessary(enabledFeatures)
-	enableTrafficLog := app.isFeatureEnabled(enabledFeatures, config.FeatureTrafficLog)
-	go app.toggleTrafficLog(enableTrafficLog)
 }
 
 // startFeaturesService starts a new features service that dispatches features to any relevant listeners.
@@ -364,29 +343,25 @@ func (app *App) beforeStart(listenAddr string) {
 		app.Exit(nil)
 		os.Exit(0)
 	}
-	app.AddExitFunc("stopping loconf scanner", LoconfScanner(app.settings, app.configDir, 4*time.Hour, app.IsProUser, func() string {
-		return app.AddToken("/img/lantern_logo.png")
-	}))
+	app.AddExitFunc("stopping loconf scanner", LoconfScanner(app.settings, app.configDir, 4*time.Hour,
+		func() (bool, bool) { return app.IsProUser(context.Background()) }, func() string {
+			return "/img/lantern_logo.png"
+		}))
 	app.AddExitFunc("stopping notifier", notifier.NotificationsLoop(app.analyticsSession))
-}
-
-func (app *App) isFeatureEnabled(features map[string]bool, feature string) bool {
-	val, ok := features[feature]
-	return ok && val
 }
 
 // Connect turns on proxying
 func (app *App) Connect() {
 	app.analyticsSession.Event("systray-menu", "connect")
 	ops.Begin("connect").End()
-	app.settings.setBool(SNDisconnected, false)
+	app.settings.SetDisconnected(false)
 }
 
 // Disconnect turns off proxying
 func (app *App) Disconnect() {
 	app.analyticsSession.Event("systray-menu", "disconnect")
 	ops.Begin("disconnect").End()
-	app.settings.setBool(SNDisconnected, true)
+	app.settings.SetDisconnected(true)
 }
 
 // GetLanguage returns the user language
@@ -407,7 +382,7 @@ func (app *App) SetLanguage(lang string) {
 
 // OnSettingChange sets a callback cb to get called when attr is changed from server.
 // When calling multiple times for same attr, only the last one takes effect.
-func (app *App) OnSettingChange(attr SettingName, cb func(interface{})) {
+func (app *App) OnSettingChange(attr settings.SettingName, cb func(interface{})) {
 	app.settings.OnChange(attr, cb)
 }
 
@@ -430,13 +405,13 @@ func (app *App) afterStart(cl *client.Client) {
 		app.SysProxyOff()
 	})
 	app.AddExitFunc("flushing to opentelemetry", otel.Stop)
-	if addr, ok := client.Addr(6 * time.Second); ok {
-		app.settings.setString(SNAddr, addr)
+	if addr, ok := flashlightClient.Addr(6 * time.Second); ok {
+		app.settings.SetAddr(addr.(string))
 	} else {
 		log.Errorf("Couldn't retrieve HTTP proxy addr in time")
 	}
-	if socksAddr, ok := client.Socks5Addr(6 * time.Second); ok {
-		app.settings.setString(SNSOCKSAddr, socksAddr)
+	if socksAddr, ok := flashlightClient.Socks5Addr(6 * time.Second); ok {
+		app.settings.SetSOCKSAddr(socksAddr.(string))
 	} else {
 		log.Errorf("Couldn't retrieve SOCKS proxy addr in time")
 	}
@@ -471,30 +446,6 @@ func (app *App) onProxiesUpdate(proxies []bandit.Dialer, src config.Source) {
 	if src == config.Fetched {
 		atomic.StoreInt32(&app.fetchedProxiesConfig, 1)
 	}
-	app.trafficLogLock.Lock()
-	app.proxiesLock.Lock()
-	app.proxies = proxies
-	if app.trafficLog != nil {
-		proxyAddresses := []string{}
-		for _, p := range proxies {
-			proxyAddresses = append(proxyAddresses, p.Addr())
-		}
-		if err := app.trafficLog.UpdateAddresses(proxyAddresses); err != nil {
-			log.Errorf("failed to update traffic log addresses: %v", err)
-		}
-	}
-	app.proxiesLock.Unlock()
-	app.trafficLogLock.Unlock()
-}
-
-// getProxies returns the currently configured proxies. State-changing methods on these dialers
-// should not be called. In short, the elements of this slice should be treated as read-only.
-func (app *App) getProxies() []bandit.Dialer {
-	app.proxiesLock.RLock()
-	copied := make([]bandit.Dialer, len(app.proxies))
-	copy(copied, app.proxies)
-	app.proxiesLock.RUnlock()
-	return copied
 }
 
 // AddExitFunc adds a function to be called before the application exits.
@@ -602,7 +553,7 @@ func (app *App) exitOnFatal(err error) {
 
 // IsPro indicates whether or not the app is pro
 func (app *App) IsPro() bool {
-	isPro, _ := app.isProUserFast()
+	isPro, _ := app.isProUserFast(context.Background())
 	return isPro
 }
 
@@ -610,12 +561,13 @@ func (app *App) IsPro() bool {
 func (app *App) ReferralCode(uc common.UserConfig) (string, error) {
 	referralCode := app.referralCode
 	if referralCode == "" {
-		resp, err := app.proClient.UserData(uc)
+		resp, err := app.proClient.UserData(context.Background())
 		if err != nil {
-			return "", err
+			return "", errors.New("error fetching user data: %v", err)
 		}
-		app.SetReferralCode(resp.Code)
-		return resp.Code, nil
+
+		app.SetReferralCode(resp.User.Code)
+		return resp.User.Code, nil
 	}
 	return referralCode, nil
 }
@@ -654,35 +606,6 @@ func ShouldReportToSentry() bool {
 	return !uicommon.IsDevEnvironment()
 }
 
-// OnTrayShow indicates the user has selected to show lantern from the tray.
-func (app *App) OnTrayShow() {
-	app.analyticsSession.Event("systray-menu", "show")
-}
-
-// OnTrayUpgrade indicates the user has selected to upgrade lantern from the tray.
-func (app *App) OnTrayUpgrade() {
-	app.analyticsSession.Event("systray-menu", "upgrade")
-}
-
-// PlansURL returns the URL for accessing the checkout/plans page directly.
-func (app *App) PlansURL() string {
-	return "#/plans"
-}
-
-// AdTrackURL returns the URL for adding tracking on injected ads.
-func (app *App) AdTrackURL() string {
-	return "/ad_track"
-}
-
-// AddToken adds our secure token to a given request path.
-func (app *App) AddToken(path string) string {
-	return path
-}
-
-func (app *App) Settings() *Settings {
-	return app.settings
-}
-
 // GetTranslations accesses translations with the given filename
 func (app *App) GetTranslations(filename string) ([]byte, error) {
 	log.Tracef("Accessing translations %v", filename)
@@ -694,5 +617,9 @@ func (app *App) GetTranslations(filename string) ([]byte, error) {
 	if err != nil {
 		return nil, fmt.Errorf("could not get traslation for file name: %v, %w", filename, err)
 	}
-	return ioutil.ReadAll(f)
+	return io.ReadAll(f)
+}
+
+func (app *App) Settings() *settings.Settings {
+	return app.settings
 }
