@@ -1,15 +1,15 @@
 package app
 
 import (
-	"io"
+	"context"
 	"math"
 	"strconv"
 
-	"github.com/getlantern/flashlight/v7/bandit"
+	"github.com/getlantern/lantern-client/desktop/settings"
+	"github.com/getlantern/lantern-client/internalsdk/pro"
+
 	"github.com/getlantern/flashlight/v7/common"
 	"github.com/getlantern/flashlight/v7/issue"
-	"github.com/getlantern/flashlight/v7/pro"
-
 	"github.com/getlantern/flashlight/v7/util"
 	"github.com/getlantern/osversion"
 )
@@ -23,9 +23,8 @@ var (
 )
 
 type issueReporter struct {
-	settings           *Settings
-	getCapturedPackets func(io.Writer) error
-	getProxies         func() []bandit.Dialer
+	settings  *settings.Settings
+	proClient pro.ProClient
 }
 
 type issueMessage struct {
@@ -44,27 +43,16 @@ type issueMessage struct {
 
 // newIssueReporter creates a new issue reporter that can be used to send issue reports
 // to the Lantern team.
-func newIssueReporter(settings *Settings, getCapturedPackets func(io.Writer) error,
-	getProxies func() []bandit.Dialer) *issueReporter {
+func newIssueReporter(app *App) *issueReporter {
 	return &issueReporter{
-		settings:           settings,
-		getCapturedPackets: getCapturedPackets,
-		getProxies:         getProxies,
+		proClient: app.proClient,
+		settings:  app.settings,
 	}
 }
 
 // sendIssueReport creates an issue report from the given UI message and submits it to
 // lantern-cloud/issue service, which is then forwarded to the ticket system via API
 func (reporter *issueReporter) sendIssueReport(msg *issueMessage) error {
-
-	if msg.RunDiagnostics {
-		var err error
-		msg.DiagnosticsYAML, msg.ProxyCapture, err = reporter.runDiagnostics()
-		if err != nil {
-			log.Errorf("error running diagnostics: %v", err)
-		}
-	}
-
 	settings := reporter.settings
 	uc := common.NewUserConfigData(
 		common.DefaultAppName,
@@ -80,7 +68,8 @@ func (reporter *issueReporter) sendIssueReport(msg *issueMessage) error {
 		return err
 	}
 	subscriptionLevel := "free"
-	if isPro, _ := pro.IsProUser(settings); isPro {
+	ctx := context.Background()
+	if isPro, _ := IsProUser(ctx, reporter.proClient, settings); isPro {
 		subscriptionLevel = "pro"
 	}
 	var osVersion string
