@@ -31,7 +31,7 @@ class AccountMenu extends StatelessWidget {
           borderRadius: BorderRadius.only(
               topLeft: Radius.circular(16), topRight: Radius.circular(16))),
       builder: (context) {
-        return FollowUs();
+        return const FollowUs();
       },
     );
   }
@@ -62,15 +62,14 @@ class AccountMenu extends StatelessWidget {
                 )
               : const SizedBox(),
         ),
-      if (Platform.isAndroid)
-        ListItemFactory.settingsItem(
-          key: AppKeys.upgrade_lantern_pro,
-          icon: ImagePaths.pro_icon_black,
-          content: 'Upgrade to Lantern Pro'.i18n,
-          onTap: () {
-            upgradeToLanternPro(context);
-          },
-        ),
+      ListItemFactory.settingsItem(
+        key: AppKeys.upgrade_lantern_pro,
+        icon: ImagePaths.pro_icon_black,
+        content: 'Upgrade to Lantern Pro'.i18n,
+        onTap: () {
+          upgradeToLanternPro(context);
+        },
+      ),
       ListItemFactory.settingsItem(
         icon: ImagePaths.star,
         content: 'Invite Friends'.i18n,
