@@ -1,5 +1,4 @@
 // ignore_for_file: use_build_context_synchronously
-
 import 'package:email_validator/email_validator.dart';
 import 'package:lantern/common/common.dart';
 import 'package:lantern/common/ui/app_loading_dialog.dart';
@@ -88,10 +87,6 @@ class _ReportIssueState extends State<ReportIssue> {
                     initialValue: emailAddress,
                     controller: emailController,
                     autovalidateMode: AutovalidateMode.disabled,
-                    contentPadding: const EdgeInsetsDirectional.only(
-                      top: 8.0,
-                      bottom: 8.0,
-                    ),
                     label: 'email'.i18n,
                     onChanged: (value) {
                       setState(() {});
@@ -138,6 +133,12 @@ class _ReportIssueState extends State<ReportIssue> {
                             issueController.text = newValue!;
                           });
                         },
+                        padding: isDesktop()
+                            ? const EdgeInsetsDirectional.only(
+                                top: 8,
+                                bottom: 8,
+                              )
+                            : const EdgeInsetsDirectional.all(0),
                         items: <String>[
                           'cannot_access_blocked_sites'.i18n,
                           'cannot_complete_purchase'.i18n,
@@ -161,8 +162,9 @@ class _ReportIssueState extends State<ReportIssue> {
                 child: CTextField(
                   tooltipMessage: 'report_description'.i18n,
                   controller: descController,
-                  contentPadding: const EdgeInsetsDirectional.all(8.0),
-                  label: '',
+                  contentPadding: isDesktop()
+                      ? const EdgeInsetsDirectional.all(16.0)
+                      : const EdgeInsetsDirectional.all(8.0),
                   hintText: 'issue_description'.i18n,
                   autovalidateMode: AutovalidateMode.disabled,
                   maxLines: 8,
