@@ -7,7 +7,7 @@ import 'package:lantern/common/common.dart';
 class CTextField extends StatefulWidget {
   late final CustomTextEditingController controller;
   late final String? initialValue;
-  late final dynamic label;
+  late final dynamic? label;
   late final String? helperText;
   late final String? hintText;
   late final Widget? prefixIcon;
@@ -36,7 +36,7 @@ class CTextField extends StatefulWidget {
   CTextField({
     required this.controller,
     this.initialValue,
-    required this.label,
+    this.label,
     this.helperText,
     this.hintText,
     this.prefixIcon,
@@ -154,7 +154,15 @@ class _CTextFieldState extends State<CTextField> {
                     widget.textCapitalization ?? TextCapitalization.none,
                 decoration: InputDecoration(
                   contentPadding: widget.contentPadding ??
-                      const EdgeInsetsDirectional.all(0),
+                      (isDesktop()
+                          ? const EdgeInsetsDirectional.only(
+                              top: 24,
+                              bottom: 24,
+                            )
+                          : const EdgeInsetsDirectional.only(
+                              top: 8,
+                              bottom: 8,
+                            )),
                   isDense: true,
                   floatingLabelBehavior: FloatingLabelBehavior.never,
                   // we handle floating labels using our custom method below
@@ -204,29 +212,30 @@ class _CTextFieldState extends State<CTextField> {
           ),
         ),
         // * Label
-        (widget.label is String)
-            ? Container(
-                margin: const EdgeInsetsDirectional.only(start: 11),
-                padding: EdgeInsetsDirectional.only(
-                  start: hasFocus ? 2 : 0,
-                  end: hasFocus ? 2 : 0,
-                ),
-                color: white,
-                child: !hasFocus && widget.controller.value.text.isEmpty
-                    ? Container()
-                    : CText(
-                        widget.label,
-                        style: CTextStyle(
-                          fontSize: 12,
-                          lineHeight: 12,
-                          color: fieldKey.currentState?.mounted == true &&
-                                  fieldKey.currentState?.hasError == true
-                              ? indicatorRed
-                              : blue4,
+        if (widget.label != null)
+          (widget.label is String)
+              ? Container(
+                  margin: const EdgeInsetsDirectional.only(start: 11),
+                  padding: EdgeInsetsDirectional.only(
+                    start: hasFocus ? 2 : 0,
+                    end: hasFocus ? 2 : 0,
+                  ),
+                  color: white,
+                  child: !hasFocus && widget.controller.value.text.isEmpty
+                      ? Container()
+                      : CText(
+                          widget.label,
+                          style: CTextStyle(
+                            fontSize: 12,
+                            lineHeight: 12,
+                            color: fieldKey.currentState?.mounted == true &&
+                                    fieldKey.currentState?.hasError == true
+                                ? indicatorRed
+                                : blue4,
+                          ),
                         ),
-                      ),
-              )
-            : widget.label,
+                )
+              : widget.label,
       ],
     );
   }
