@@ -98,13 +98,14 @@ const Plan$json = {
     {'1': 'bestValue', '3': 3, '4': 1, '5': 8, '10': 'bestValue'},
     {'1': 'usdPrice', '3': 4, '4': 1, '5': 3, '10': 'usdPrice'},
     {'1': 'price', '3': 5, '4': 3, '5': 11, '6': '.Plan.PriceEntry', '10': 'price'},
-    {'1': 'totalCostBilledOneTime', '3': 6, '4': 1, '5': 9, '10': 'totalCostBilledOneTime'},
-    {'1': 'oneMonthCost', '3': 7, '4': 1, '5': 9, '10': 'oneMonthCost'},
-    {'1': 'totalCost', '3': 8, '4': 1, '5': 9, '10': 'totalCost'},
-    {'1': 'formattedBonus', '3': 9, '4': 1, '5': 9, '10': 'formattedBonus'},
-    {'1': 'renewalText', '3': 10, '4': 1, '5': 9, '10': 'renewalText'},
+    {'1': 'expectedMonthlyPrice', '3': 6, '4': 3, '5': 11, '6': '.Plan.ExpectedMonthlyPriceEntry', '10': 'expectedMonthlyPrice'},
+    {'1': 'totalCostBilledOneTime', '3': 7, '4': 1, '5': 9, '10': 'totalCostBilledOneTime'},
+    {'1': 'oneMonthCost', '3': 8, '4': 1, '5': 9, '10': 'oneMonthCost'},
+    {'1': 'totalCost', '3': 9, '4': 1, '5': 9, '10': 'totalCost'},
+    {'1': 'formattedBonus', '3': 10, '4': 1, '5': 9, '10': 'formattedBonus'},
+    {'1': 'renewalText', '3': 11, '4': 1, '5': 9, '10': 'renewalText'},
   ],
-  '3': [Plan_PriceEntry$json],
+  '3': [Plan_PriceEntry$json, Plan_ExpectedMonthlyPriceEntry$json],
 };
 
 @$core.Deprecated('Use planDescriptor instead')
@@ -117,30 +118,57 @@ const Plan_PriceEntry$json = {
   '7': {'7': true},
 };
 
+@$core.Deprecated('Use planDescriptor instead')
+const Plan_ExpectedMonthlyPriceEntry$json = {
+  '1': 'ExpectedMonthlyPriceEntry',
+  '2': [
+    {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
+    {'1': 'value', '3': 2, '4': 1, '5': 3, '10': 'value'},
+  ],
+  '7': {'7': true},
+};
+
 /// Descriptor for `Plan`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List planDescriptor = $convert.base64Decode(
     'CgRQbGFuEg4KAmlkGAEgASgJUgJpZBIgCgtkZXNjcmlwdGlvbhgCIAEoCVILZGVzY3JpcHRpb2'
     '4SHAoJYmVzdFZhbHVlGAMgASgIUgliZXN0VmFsdWUSGgoIdXNkUHJpY2UYBCABKANSCHVzZFBy'
-    'aWNlEiYKBXByaWNlGAUgAygLMhAuUGxhbi5QcmljZUVudHJ5UgVwcmljZRI2ChZ0b3RhbENvc3'
-    'RCaWxsZWRPbmVUaW1lGAYgASgJUhZ0b3RhbENvc3RCaWxsZWRPbmVUaW1lEiIKDG9uZU1vbnRo'
-    'Q29zdBgHIAEoCVIMb25lTW9udGhDb3N0EhwKCXRvdGFsQ29zdBgIIAEoCVIJdG90YWxDb3N0Ei'
-    'YKDmZvcm1hdHRlZEJvbnVzGAkgASgJUg5mb3JtYXR0ZWRCb251cxIgCgtyZW5ld2FsVGV4dBgK'
-    'IAEoCVILcmVuZXdhbFRleHQaOAoKUHJpY2VFbnRyeRIQCgNrZXkYASABKAlSA2tleRIUCgV2YW'
-    'x1ZRgCIAEoA1IFdmFsdWU6AjgB');
+    'aWNlEiYKBXByaWNlGAUgAygLMhAuUGxhbi5QcmljZUVudHJ5UgVwcmljZRJTChRleHBlY3RlZE'
+    '1vbnRobHlQcmljZRgGIAMoCzIfLlBsYW4uRXhwZWN0ZWRNb250aGx5UHJpY2VFbnRyeVIUZXhw'
+    'ZWN0ZWRNb250aGx5UHJpY2USNgoWdG90YWxDb3N0QmlsbGVkT25lVGltZRgHIAEoCVIWdG90YW'
+    'xDb3N0QmlsbGVkT25lVGltZRIiCgxvbmVNb250aENvc3QYCCABKAlSDG9uZU1vbnRoQ29zdBIc'
+    'Cgl0b3RhbENvc3QYCSABKAlSCXRvdGFsQ29zdBImCg5mb3JtYXR0ZWRCb251cxgKIAEoCVIOZm'
+    '9ybWF0dGVkQm9udXMSIAoLcmVuZXdhbFRleHQYCyABKAlSC3JlbmV3YWxUZXh0GjgKClByaWNl'
+    'RW50cnkSEAoDa2V5GAEgASgJUgNrZXkSFAoFdmFsdWUYAiABKANSBXZhbHVlOgI4ARpHChlFeH'
+    'BlY3RlZE1vbnRobHlQcmljZUVudHJ5EhAKA2tleRgBIAEoCVIDa2V5EhQKBXZhbHVlGAIgASgD'
+    'UgV2YWx1ZToCOAE=');
 
 @$core.Deprecated('Use paymentProvidersDescriptor instead')
 const PaymentProviders$json = {
   '1': 'PaymentProviders',
   '2': [
     {'1': 'name', '3': 1, '4': 1, '5': 9, '10': 'name'},
-    {'1': 'logoUrls', '3': 2, '4': 3, '5': 9, '10': 'logoUrls'},
+    {'1': 'data', '3': 2, '4': 3, '5': 11, '6': '.PaymentProviders.DataEntry', '10': 'data'},
+    {'1': 'logoUrls', '3': 3, '4': 3, '5': 9, '10': 'logoUrls'},
   ],
+  '3': [PaymentProviders_DataEntry$json],
+};
+
+@$core.Deprecated('Use paymentProvidersDescriptor instead')
+const PaymentProviders_DataEntry$json = {
+  '1': 'DataEntry',
+  '2': [
+    {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
+    {'1': 'value', '3': 2, '4': 1, '5': 9, '10': 'value'},
+  ],
+  '7': {'7': true},
 };
 
 /// Descriptor for `PaymentProviders`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List paymentProvidersDescriptor = $convert.base64Decode(
-    'ChBQYXltZW50UHJvdmlkZXJzEhIKBG5hbWUYASABKAlSBG5hbWUSGgoIbG9nb1VybHMYAiADKA'
-    'lSCGxvZ29Vcmxz');
+    'ChBQYXltZW50UHJvdmlkZXJzEhIKBG5hbWUYASABKAlSBG5hbWUSLwoEZGF0YRgCIAMoCzIbLl'
+    'BheW1lbnRQcm92aWRlcnMuRGF0YUVudHJ5UgRkYXRhEhoKCGxvZ29VcmxzGAMgAygJUghsb2dv'
+    'VXJscxo3CglEYXRhRW50cnkSEAoDa2V5GAEgASgJUgNrZXkSFAoFdmFsdWUYAiABKAlSBXZhbH'
+    'VlOgI4AQ==');
 
 @$core.Deprecated('Use paymentMethodDescriptor instead')
 const PaymentMethod$json = {
@@ -164,14 +192,15 @@ const User$json = {
     {'1': 'email', '3': 2, '4': 1, '5': 9, '10': 'email'},
     {'1': 'telephone', '3': 3, '4': 1, '5': 9, '10': 'telephone'},
     {'1': 'userStatus', '3': 4, '4': 1, '5': 9, '10': 'userStatus'},
-    {'1': 'locale', '3': 5, '4': 1, '5': 9, '10': 'locale'},
-    {'1': 'expiration', '3': 6, '4': 1, '5': 3, '10': 'expiration'},
-    {'1': 'devices', '3': 7, '4': 3, '5': 11, '6': '.Device', '10': 'devices'},
-    {'1': 'code', '3': 8, '4': 1, '5': 9, '10': 'code'},
-    {'1': 'expireAt', '3': 9, '4': 1, '5': 3, '10': 'expireAt'},
-    {'1': 'referral', '3': 10, '4': 1, '5': 9, '10': 'referral'},
-    {'1': 'token', '3': 11, '4': 1, '5': 9, '10': 'token'},
-    {'1': 'yinbiEnabled', '3': 12, '4': 1, '5': 8, '10': 'yinbiEnabled'},
+    {'1': 'userLevel', '3': 5, '4': 1, '5': 9, '10': 'userLevel'},
+    {'1': 'locale', '3': 6, '4': 1, '5': 9, '10': 'locale'},
+    {'1': 'expiration', '3': 7, '4': 1, '5': 3, '10': 'expiration'},
+    {'1': 'devices', '3': 8, '4': 3, '5': 11, '6': '.Device', '10': 'devices'},
+    {'1': 'code', '3': 9, '4': 1, '5': 9, '10': 'code'},
+    {'1': 'expireAt', '3': 10, '4': 1, '5': 3, '10': 'expireAt'},
+    {'1': 'referral', '3': 11, '4': 1, '5': 9, '10': 'referral'},
+    {'1': 'token', '3': 12, '4': 1, '5': 9, '10': 'token'},
+    {'1': 'yinbiEnabled', '3': 13, '4': 1, '5': 8, '10': 'yinbiEnabled'},
   ],
 };
 
@@ -179,15 +208,15 @@ const User$json = {
 final $typed_data.Uint8List userDescriptor = $convert.base64Decode(
     'CgRVc2VyEhYKBnVzZXJJZBgBIAEoA1IGdXNlcklkEhQKBWVtYWlsGAIgASgJUgVlbWFpbBIcCg'
     'l0ZWxlcGhvbmUYAyABKAlSCXRlbGVwaG9uZRIeCgp1c2VyU3RhdHVzGAQgASgJUgp1c2VyU3Rh'
-    'dHVzEhYKBmxvY2FsZRgFIAEoCVIGbG9jYWxlEh4KCmV4cGlyYXRpb24YBiABKANSCmV4cGlyYX'
-    'Rpb24SIQoHZGV2aWNlcxgHIAMoCzIHLkRldmljZVIHZGV2aWNlcxISCgRjb2RlGAggASgJUgRj'
-    'b2RlEhoKCGV4cGlyZUF0GAkgASgDUghleHBpcmVBdBIaCghyZWZlcnJhbBgKIAEoCVIIcmVmZX'
-    'JyYWwSFAoFdG9rZW4YCyABKAlSBXRva2VuEiIKDHlpbmJpRW5hYmxlZBgMIAEoCFIMeWluYmlF'
-    'bmFibGVk');
+    'dHVzEhwKCXVzZXJMZXZlbBgFIAEoCVIJdXNlckxldmVsEhYKBmxvY2FsZRgGIAEoCVIGbG9jYW'
+    'xlEh4KCmV4cGlyYXRpb24YByABKANSCmV4cGlyYXRpb24SIQoHZGV2aWNlcxgIIAMoCzIHLkRl'
+    'dmljZVIHZGV2aWNlcxISCgRjb2RlGAkgASgJUgRjb2RlEhoKCGV4cGlyZUF0GAogASgDUghleH'
+    'BpcmVBdBIaCghyZWZlcnJhbBgLIAEoCVIIcmVmZXJyYWwSFAoFdG9rZW4YDCABKAlSBXRva2Vu'
+    'EiIKDHlpbmJpRW5hYmxlZBgNIAEoCFIMeWluYmlFbmFibGVk');
 
-@$core.Deprecated('Use aPIResponseDescriptor instead')
-const APIResponse$json = {
-  '1': 'APIResponse',
+@$core.Deprecated('Use baseResponseDescriptor instead')
+const BaseResponse$json = {
+  '1': 'BaseResponse',
   '2': [
     {'1': 'status', '3': 1, '4': 1, '5': 9, '10': 'status'},
     {'1': 'error', '3': 2, '4': 1, '5': 9, '10': 'error'},
@@ -195,10 +224,51 @@ const APIResponse$json = {
   ],
 };
 
-/// Descriptor for `APIResponse`. Decode as a `google.protobuf.DescriptorProto`.
-final $typed_data.Uint8List aPIResponseDescriptor = $convert.base64Decode(
-    'CgtBUElSZXNwb25zZRIWCgZzdGF0dXMYASABKAlSBnN0YXR1cxIUCgVlcnJvchgCIAEoCVIFZX'
-    'Jyb3ISGAoHZXJyb3JJZBgDIAEoCVIHZXJyb3JJZA==');
+/// Descriptor for `BaseResponse`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List baseResponseDescriptor = $convert.base64Decode(
+    'CgxCYXNlUmVzcG9uc2USFgoGc3RhdHVzGAEgASgJUgZzdGF0dXMSFAoFZXJyb3IYAiABKAlSBW'
+    'Vycm9yEhgKB2Vycm9ySWQYAyABKAlSB2Vycm9ySWQ=');
+
+@$core.Deprecated('Use paymentRedirectRequestDescriptor instead')
+const PaymentRedirectRequest$json = {
+  '1': 'PaymentRedirectRequest',
+  '2': [
+    {'1': 'plan', '3': 1, '4': 1, '5': 9, '10': 'plan'},
+    {'1': 'provider', '3': 2, '4': 1, '5': 9, '10': 'provider'},
+    {'1': 'currency', '3': 3, '4': 1, '5': 9, '10': 'currency'},
+    {'1': 'email', '3': 4, '4': 1, '5': 9, '10': 'email'},
+    {'1': 'deviceName', '3': 5, '4': 1, '5': 9, '10': 'deviceName'},
+    {'1': 'countryCode', '3': 6, '4': 1, '5': 9, '10': 'countryCode'},
+    {'1': 'locale', '3': 7, '4': 1, '5': 9, '10': 'locale'},
+  ],
+};
+
+/// Descriptor for `PaymentRedirectRequest`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List paymentRedirectRequestDescriptor = $convert.base64Decode(
+    'ChZQYXltZW50UmVkaXJlY3RSZXF1ZXN0EhIKBHBsYW4YASABKAlSBHBsYW4SGgoIcHJvdmlkZX'
+    'IYAiABKAlSCHByb3ZpZGVyEhoKCGN1cnJlbmN5GAMgASgJUghjdXJyZW5jeRIUCgVlbWFpbBgE'
+    'IAEoCVIFZW1haWwSHgoKZGV2aWNlTmFtZRgFIAEoCVIKZGV2aWNlTmFtZRIgCgtjb3VudHJ5Q2'
+    '9kZRgGIAEoCVILY291bnRyeUNvZGUSFgoGbG9jYWxlGAcgASgJUgZsb2NhbGU=');
+
+@$core.Deprecated('Use redeemResellerCodeRequestDescriptor instead')
+const RedeemResellerCodeRequest$json = {
+  '1': 'RedeemResellerCodeRequest',
+  '2': [
+    {'1': 'email', '3': 1, '4': 1, '5': 9, '10': 'email'},
+    {'1': 'resellerCode', '3': 2, '4': 1, '5': 9, '10': 'resellerCode'},
+    {'1': 'deviceName', '3': 3, '4': 1, '5': 9, '10': 'deviceName'},
+    {'1': 'currency', '3': 4, '4': 1, '5': 9, '10': 'currency'},
+    {'1': 'idempotencyKey', '3': 5, '4': 1, '5': 9, '10': 'idempotencyKey'},
+    {'1': 'provider', '3': 6, '4': 1, '5': 9, '10': 'provider'},
+  ],
+};
+
+/// Descriptor for `RedeemResellerCodeRequest`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List redeemResellerCodeRequestDescriptor = $convert.base64Decode(
+    'ChlSZWRlZW1SZXNlbGxlckNvZGVSZXF1ZXN0EhQKBWVtYWlsGAEgASgJUgVlbWFpbBIiCgxyZX'
+    'NlbGxlckNvZGUYAiABKAlSDHJlc2VsbGVyQ29kZRIeCgpkZXZpY2VOYW1lGAMgASgJUgpkZXZp'
+    'Y2VOYW1lEhoKCGN1cnJlbmN5GAQgASgJUghjdXJyZW5jeRImCg5pZGVtcG90ZW5jeUtleRgFIA'
+    'EoCVIOaWRlbXBvdGVuY3lLZXkSGgoIcHJvdmlkZXIYBiABKAlSCHByb3ZpZGVy');
 
 @$core.Deprecated('Use paymentRedirectResponseDescriptor instead')
 const PaymentRedirectResponse$json = {
